feat: verify if folder is empty before trying to run script
